card-replacement-agent/pyproject.toml
[project]
name = "mini-bcs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.2",
    "openai-agents>=0.3.0",
    "gradio>=4.44.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.1.1",
    "tinydb>=4.8.2",
    "fastapi-mcp>=0.4.0",
    "psycopg2-binary>=2.9.9",
]

[tool.uv.sources]
